+++ package.json
@@ -13,6 +13,7 @@
     "@testing-library/user-event": "^7.2.1",
     "axios": "^0.19.2",
     "cloudinary-core": "^2.11.1",
+    "cloudinary-react": "^1.6.6",
     "formik": "^2.1.4",
     "lodash": "^4.17.19",
     "node-sass": "^4.14.1",
@@ -20,7 +21,7 @@
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1",
+    "react-scripts": "^3.4.3",
     "react-transition-group": "^4.4.1",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
