+++ public/index.html
@@ -43,6 +43,10 @@
       href="https://fonts.googleapis.com/css2?family=Lato&family=Open+Sans:wght@400;700&display=swap"
       rel="stylesheet"
     />
+    <script
+      src="https://unpkg.com/cloudinary-core/cloudinary-core-shrinkwrap.min.js"
+      type="text/javascript"
+    ></script>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
